requirements.txt
Django
Pillow
tensorflow
keras
numpy
scipy
opencv-python
scikit-image
matplotlib
scikit-learn
scikeras
mysqlclient